+++ adminalerts/forms.py
@@ -9,16 +9,13 @@ class AdminAlertForm(forms.ModelForm):
 
     class Meta:
         model = AdminAlert
-        fields = ['message', 'user', 'date_expire', 'description', 'active']
+        fields = ['message', 'date_expire', 'description', 'active']
 
     def __init__(self, current_user=None, *args, **kwargs):
         """Override for form initialization"""
         super(AdminAlertForm, self).__init__(*args, **kwargs)
 
         self.current_user = current_user
-
-        # Hide user
-        self.fields['user'].widget = forms.HiddenInput()
 
         # Set date_expire properties
         # NOTE: "format" works in source but not in widget, any way to fix?
@@ -29,7 +26,6 @@ class AdminAlertForm(forms.ModelForm):
 
         # Set initial values if creating
         if not self.instance.pk:
-            self.fields['user'].initial = self.current_user
             self.fields['date_expire'].initial = timezone.now() + \
                 timezone.timedelta(days=1)
 
@@ -43,3 +39,9 @@ class AdminAlertForm(forms.ModelForm):
 
         return self.cleaned_data
 
+    def save(self, *args, **kwargs):
+        """Override of form saving function"""
+        obj = super(AdminAlertForm, self).save(commit=False)
+        obj.user = self.current_user
+        obj.save()
+        return obj